This commit refactors the DAP configuration by splitting it intoseparate files for adapters, configurations, and UI setup. This makesthe configuration more modular and easier to maintain.In addition, a new feature is added to automatically scroll certain DAPwindows to the bottom. This is useffeat(dap): Refactor DAP configuration and add auto-scroll
This commit refactors the DAP configuration by splitting it intoseparate
files for adapters, configurations, and UI setup. This makesthe
configuration more modular and easier to maintain.In addition, a new
feature is added to automatically scroll certain DAPwindows to the
bottom. This is useful for keeping track of the latestoutput in these
windows.A new keymap is also added to close the DAP UI, and a new
command isadded to print the current buffer name.